% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hline.R
\name{hline}
\alias{hline}
\title{Simple horizontal line plotly}
\usage{
hline(y, color = "red")
}
\arguments{
\item{y}{vertical coordinate of the horizontal line}

\item{color}{color of the horizontal line.}
}
\value{
plotly list type line
}
\description{
Simple horizontal line plotly
}
