% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_annotation.R
\name{ms_annotation}
\alias{ms_annotation}
\title{Simple annotation for MS values of PWD plot.}
\usage{
ms_annotation(label, y)
}
\arguments{
\item{label}{label of the annotation}

\item{y}{coordinate of the text annotation}
}
\value{
plotly list annotation
}
\description{
Simple annotation for MS values of PWD plot.
}
